requirements.txt
fastapi
sqlalchemy
redis
celery
uvicorn
python-dotenv
python-multipart
LangChain
ChromaDB
Postgres
PyMuPDF
pytesseract
PaddleOCR
python-docx